Add go build tags
