+++ docs/data-sources/packages.md
@@ -24,7 +24,7 @@ data "tanzu-mission-control_packages" "read_cluster_packages" {
 
   scope {
     cluster {
-      cluster_name            = "testcluster" # Required
+      name            = "testcluster" # Required
       provisioner_name        = "attached"    # Default: attached
       management_cluster_name = "attached"    # Default: attached
     }
@@ -53,7 +53,7 @@ data "tanzu-mission-control_packages" "read_cluster_packages" {
 
 Optional:
 
-- `cluster` (Block List, Max: 1) The schema for cluster iam policy full name (see [below for nested schema](#nestedblock--scope--cluster))
+- `cluster` (Block List, Max: 1) The schema for cluster full name (see [below for nested schema](#nestedblock--scope--cluster))
 
 <a id="nestedblock--scope--cluster"></a>
 ### Nested Schema for `scope.cluster`
