+++ docs/resources/package_install.md
@@ -30,7 +30,7 @@ resource "tanzu-mission-control_package_install" "create_package_install" {
 
     scope {
         cluster {
-        cluster_name            = "testcluster" # Required
+        name            = "testcluster" # Required
         provisioner_name        = "attached"    # Default: attached
         management_cluster_name = "attached"    # Default: attached
         }
@@ -73,7 +73,7 @@ resource "tanzu-mission-control_package_install" "create_package_install" {
 
 Optional:
 
-- `cluster` (Block List, Max: 1) The schema for cluster iam policy full name (see [below for nested schema](#nestedblock--scope--cluster))
+- `cluster` (Block List, Max: 1) The schema for cluster full name (see [below for nested schema](#nestedblock--scope--cluster))
 
 <a id="nestedblock--scope--cluster"></a>
 ### Nested Schema for `scope.cluster`
