+++ examples/data-sources/packageinstall/data-source.tf
@@ -6,7 +6,7 @@ data "tanzu-mission-control_package_install" "read_package_install" {
 
     scope {
         cluster {
-        cluster_name            = "testcluster" # Required
+        name            = "testcluster" # Required
         provisioner_name        = "attached"    # Default: attached
         management_cluster_name = "attached"    # Default: attached
         }
