+++ examples/data-sources/packages/data-source.tf
@@ -4,7 +4,7 @@ data "tanzu-mission-control_packages" "read_cluster_packages" {
 
   scope {
     cluster {
-      cluster_name            = "testcluster" # Required
+      name            = "testcluster" # Required
       provisioner_name        = "attached"    # Default: attached
       management_cluster_name = "attached"    # Default: attached
     }
