+++ examples/resources/tanzupackagerepository/resource.tf
@@ -4,7 +4,7 @@ resource "tanzu-mission-control_package_repository" "create_cluster_pkg_reposito
 
   scope {
     cluster {
-      cluster_name            = "testcluster" # Required
+      name            = "testcluster" # Required
       provisioner_name        = "attached"    # Default: attached
       management_cluster_name = "attached"    # Default: attached
     }
