+++ internal/resources/packages/packages_mock_test.go
@@ -32,7 +32,6 @@ const (
 	packages             = "packages"
 )
 
-// nolint: unparam
 func bodyInspectingResponder(t *testing.T, expectedContent interface{}, successResponse int, successResponseBody interface{}) httpmock.Responder {
 	return func(r *http.Request) (*http.Response, error) {
 		successFunc := func() (*http.Response, error) {
