+++ internal/resources/tanzupackageinstall/package_install_resource_test.go
@@ -1,3 +1,6 @@
+//go:build packageinstall
+// +build packageinstall
+
 /*
 Copyright © 2023 VMware, Inc. All Rights Reserved.
 SPDX-License-Identifier: MPL-2.0
@@ -6,17 +9,14 @@ SPDX-License-Identifier: MPL-2.0
 package tanzupackageinstall
 
 import (
-	"context"
 	"fmt"
 	"log"
 	"os"
 	"strings"
 	"testing"
 
-	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/acctest"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
 	"github.com/pkg/errors"
 
@@ -25,36 +25,8 @@ import (
 	packageinstallclustermodel "github.com/vmware/terraform-provider-tanzu-mission-control/internal/models/tanzupackageinstall"
 	commonscope "github.com/vmware/terraform-provider-tanzu-mission-control/internal/resources/common/scope"
 	packageinstallscope "github.com/vmware/terraform-provider-tanzu-mission-control/internal/resources/tanzupackageinstall/scope"
-	packagerepository "github.com/vmware/terraform-provider-tanzu-mission-control/internal/resources/tanzupackagerepository"
 	testhelper "github.com/vmware/terraform-provider-tanzu-mission-control/internal/resources/testing"
 )
-
-// nolint: gosec
-const (
-	pkgRepoResource      = packagerepository.ResourceName
-	pkgRepoResourceVar   = "test_pkg_repository"
-	pkgRepoDataSourceVar = "test_data_source_pkg_repository"
-	pkgRepoNamePrefix    = "tf-pkg-repository-test"
-
-	pkgInstallResource      = ResourceName
-	pkgInstallResourceVar   = "test_pkg_install"
-	pkgInstallDataSourceVar = "test_data_source_pkg_install"
-	pkgInstallNamePrefix    = "tf-pkg-install-test"
-	namespaceNamePrefix     = "test-pkg-install-ns"
-
-	constraints = "3.0.0-rc.1"
-)
-
-type testAcceptanceConfig struct {
-	Provider               *schema.Provider
-	PkgInstallResource     string
-	PkgInstallResourceVar  string
-	PkgInstallResourceName string
-	PkgInstallName         string
-	PkgRepoName            string
-	ScopeHelperResources   *commonscope.ScopeHelperResources
-	Namespace              string
-}
 
 func testGetDefaultAcceptanceConfig(t *testing.T) *testAcceptanceConfig {
 	return &testAcceptanceConfig{
@@ -67,14 +39,6 @@ func testGetDefaultAcceptanceConfig(t *testing.T) *testAcceptanceConfig {
 		ScopeHelperResources:   commonscope.NewScopeHelperResources(),
 		Namespace:              acctest.RandomWithPrefix(namespaceNamePrefix),
 	}
-}
-
-func getConfigureContextFunc() func(_ context.Context, d *schema.ResourceData) (interface{}, diag.Diagnostics) {
-	if _, found := os.LookupEnv("ENABLE_PKGINS_ENV_TEST"); !found {
-		return authctx.ProviderConfigureContextWithDefaultTransportForTesting
-	}
-
-	return authctx.ProviderConfigureContext
 }
 
 func getSetupConfig(config *authctx.TanzuContext) error {
