+++ internal/resources/tanzupackagerepository/data_source_repository_test.go
@@ -1,3 +1,6 @@
+//go:build packagerepository
+// +build packagerepository
+
 /*
 Copyright © 2023 VMware, Inc. All Rights Reserved.
 SPDX-License-Identifier: MPL-2.0
