+++ internal/resources/tanzupackagerepository/resource_package_repository_test.go
@@ -1,3 +1,6 @@
+//go:build packagerepository
+// +build packagerepository
+
 /*
 Copyright © 2023 VMware, Inc. All Rights Reserved.
 SPDX-License-Identifier: MPL-2.0
@@ -6,17 +9,14 @@ SPDX-License-Identifier: MPL-2.0
 package tanzupackagerepository
 
 import (
-	"context"
 	"fmt"
 	"log"
 	"os"
 	"strings"
 	"testing"
 
-	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/acctest"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
 	"github.com/pkg/errors"
 
@@ -27,29 +27,6 @@ import (
 	packagerepositoryscope "github.com/vmware/terraform-provider-tanzu-mission-control/internal/resources/tanzupackagerepository/scope"
 	testhelper "github.com/vmware/terraform-provider-tanzu-mission-control/internal/resources/testing"
 )
-
-// nolint: gosec
-const (
-	PkgRepoResource      = ResourceName
-	PkgRepoResourceVar   = "test_pkg_repository"
-	pkgRepoDataSourceVar = "test_data_source_pkg_repository"
-	pkgRepoNamePrefix    = "tf-pkg-repository-test"
-	globalRepoNamespace  = "tanzu-package-repo-global"
-
-	imageURL        = "extensions.aws-usw2.tmc-dev.cloud.vmware.com/packages/standard/repo:v2.2.0_update.2"
-	updatedImageURL = "extensions.aws-usw2.tmc-dev.cloud.vmware.com/packages/standard/repo:v2.2.0_update.1"
-)
-
-type testAcceptanceConfig struct {
-	Provider                    *schema.Provider
-	PkgRepoResource             string
-	PkgRepoResourceVar          string
-	PkgRepoResourceName         string
-	PkgRepoName                 string
-	PkgRepoDataSourceVar        string
-	PkgRepositoryDataSourceName string
-	ScopeHelperResources        *commonscope.ScopeHelperResources
-}
 
 func testGetDefaultAcceptanceConfig(t *testing.T) *testAcceptanceConfig {
 	return &testAcceptanceConfig{
@@ -62,14 +39,6 @@ func testGetDefaultAcceptanceConfig(t *testing.T) *testAcceptanceConfig {
 		PkgRepoDataSourceVar:        pkgRepoDataSourceVar,
 		PkgRepositoryDataSourceName: fmt.Sprintf("data.%s.%s", ResourceName, pkgRepoDataSourceVar),
 	}
-}
-
-func getConfigureContextFunc() func(_ context.Context, d *schema.ResourceData) (interface{}, diag.Diagnostics) {
-	if _, found := os.LookupEnv("ENABLE_PKGREPO_ENV_TEST"); !found {
-		return authctx.ProviderConfigureContextWithDefaultTransportForTesting
-	}
-
-	return authctx.ProviderConfigureContext
 }
 
 func getSetupConfig(config *authctx.TanzuContext) error {
